+++ code/graphics/vulkan/VulkanConstants.h
@@ -5,7 +5,7 @@
 namespace graphics {
 namespace vulkan {
 
-constexpr uint32_t kFramesInFlight = 3;
+constexpr uint32_t kFramesInFlight = 2;
 constexpr uint32_t kMaxBindlessTextures = 1024;
 constexpr uint32_t kModelSetsPerPool = 4096 * kFramesInFlight;